# webapp/src/main/java/ar/edu/itba/paw/webapp/controller/resource/roomify-openapi.yaml
openapi: 3.0.3
info:
  title: Roomify API - REST
  description: Official Roomify REST API documentation based on the project specifications.
  version: 1.0.0
servers:
  - url: http://localhost:8080/webapp/api
    description: Local server
...
security:
  - bearerAuth: []
paths:
...
  /reviews:
    get:
      tags: [Reviews]
      summary: Get reviews
      description: Returns a paginated list of reviews filtered by userId, roomId or roomOwnerId.
      security: []
      parameters:
        - name: userId
          in: query
          schema: { type: integer }
        - name: roomId
          in: query
          schema: { type: integer }
        - name: roomOwnerId
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: List of reviews.
          content:
            application/vnd.roomify.reviews.v1.list+json:
              schema:
                $ref: '#/components/schemas/Reviews'
    post:
      tags: [Reviews]
      summary: Leave a review
      description: Publishes a review for a room where the user stayed.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/vnd.roomify.review.v1+json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '201':
          description: Review created.
          headers:
            Location:
              description: URI of the created review.
              schema:
                type: string
                format: uri
    delete:
      tags: [Reviews]
      summary: Delete review
      description: Only the review owner can delete it.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/vnd.roomify.review.v1+json:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required: [id]
      responses:
        '204':
          description: Review deleted.
        '403':
          description: Permission denied.
...
  /images:
    post:
      tags: [Images]
      summary: Upload image
      description: Uploads a room image.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
              required: [image]
      responses:
        '201':
          description: Image uploaded.
          headers:
            Location:
              description: URI of the uploaded image.
              schema:
                type: string
                format: uri
        '401':
          description: Not authenticated.
  /images/{id}:
    get:
      tags: [Images]
      summary: Get binary image
      description: Retrieves the image binary data and returns the corresponding Content-Type.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Binary image.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Image not found.
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Review:
      type: object
      properties:
        id: { type: integer }
        reviewerName: { type: string }
        reviewerUrl: { type: string }
        roomId: { type: integer }
        roomTitle: { type: string }
        roomUrl: { type: string }
        url: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
        date: { type: string, format: date }
    Reviews:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        totalReviews: { type: integer }
        averageRating: { type: number, format: double }
    ReviewCreate:
      type: object
      required: [contactId, reviewerId, rating, comment]
      properties:
        contactId: { type: integer }
        reviewerId: { type: integer }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
...